Baseline Index Anchors Code
This was modeled off the pattern for links. This generates an HTML Anchor in the active paragraph.
The intent is to use a snippet to generate the final index by scanning all pages and ordering the index accordingly.

Format:

@Index Entry Name@(Index Parent Entry 1| Index Parent Enry 2|..)

Key:

Index Entry Name: This is the label in the index the reader will look up.
Index Parent Entry: This is an optional list of parent index entries that the anchor could be a child of.

Example:

@Fireball@(Fire Spells|Ranged Spells)

This could/should output something like:

...
Fire Spells
   Fireball    p 16
...

...
Ranged Spells
   Fireball    p 16

Entries without parents shhould be treated as top level.

Notes and other lunacy:
   Any characters valid for links except @ are permitted.
   This may not be the best set of delimiters for this purpose. Open to suggestions.
   Does not currently prevent duplicate anchors.
   Anchor target values are created by removing whitespace and lowercasing the Index Entry Name.

To Do:
  Add a Page level anchor to the \page code
  Create a Snippet to generate the actual index
  Determine a sane way to do a light index ( list of words )
  Determine a sane way to merge both Index styles